+++ piecrust/src/store/tree.rs
@@ -12,6 +12,7 @@ use std::{
 use bytecheck::CheckBytes;
 use piecrust_uplink::ContractId;
 use rkyv::{Archive, Deserialize, Serialize};
+use std::io::{self, ErrorKind, Read, Write};
 
 // There are max `2^16` pages in a 32-bit memory
 const P32_HEIGHT: usize = 8;
@@ -97,6 +98,7 @@ impl NewContractIndex {
 pub struct ContractsMerkle {
     inner_tree: Tree,
     dict: BTreeMap<u64, u64>,
+    tree_pos: TreePos,
 }
 
 impl Default for ContractsMerkle {
@@ -104,6 +106,7 @@ impl Default for ContractsMerkle {
         Self {
             inner_tree: Tree::new(),
             dict: BTreeMap::new(),
+            tree_pos: TreePos::default(),
         }
     }
 }
@@ -119,12 +122,14 @@ impl ContractsMerkle {
             Some(p) => *p,
         };
         self.inner_tree.insert(new_pos, hash);
+        self.tree_pos.insert(new_pos as u32, (hash, pos));
         new_pos
     }
 
     pub fn insert_with_int_pos(&mut self, pos: u64, int_pos: u64, hash: Hash) {
         self.dict.insert(pos, int_pos);
         self.inner_tree.insert(int_pos, hash);
+        self.tree_pos.insert(int_pos as u32, (hash, pos));
     }
 
     pub fn opening(&self, pos: u64) -> Option<TreeOpening> {
@@ -134,6 +139,14 @@ impl ContractsMerkle {
 
     pub fn root(&self) -> Ref<Hash> {
         self.inner_tree.root()
+    }
+
+    pub fn tree_pos(&self) -> &TreePos {
+        &self.tree_pos
+    }
+
+    pub fn len(&self) -> u64 {
+        self.inner_tree.len()
     }
 }
 
@@ -153,14 +166,152 @@ pub struct BaseInfo {
     pub maybe_base: Option<Hash>,
 }
 
+#[derive(Debug, Clone, Default, Archive, Deserialize, Serialize)]
+#[archive_attr(derive(CheckBytes))]
+pub struct TreePos {
+    tree_pos: BTreeMap<u32, (Hash, u64)>,
+}
+
+impl TreePos {
+    pub fn insert(&mut self, k: u32, v: (Hash, u64)) {
+        self.tree_pos.insert(k, v);
+    }
+
+    pub fn marshall<W: Write>(&self, w: &mut W) -> io::Result<()> {
+        const CHUNK_SIZE: usize = 8192;
+        const ELEM_SIZE: usize = 4 + 32 + 4;
+        let mut b = [0u8; ELEM_SIZE * CHUNK_SIZE];
+        let mut chk = 0;
+        for (k, (h, p)) in self.tree_pos.iter() {
+            let offset = chk * ELEM_SIZE;
+            b[offset..(offset + 4)].copy_from_slice(&(*k).to_le_bytes());
+            b[(offset + 4)..(offset + 36)].copy_from_slice(h.as_bytes());
+            b[(offset + 36)..(offset + 40)]
+                .copy_from_slice(&(*p as u32).to_le_bytes());
+            chk = (chk + 1) % CHUNK_SIZE;
+            if chk == 0 {
+                w.write_all(b.as_slice())?;
+            }
+        }
+        if chk != 0 {
+            w.write_all(&b[..(chk * ELEM_SIZE)])?;
+        }
+        Ok(())
+    }
+
+    fn read_bytes<R: Read, const N: usize>(r: &mut R) -> io::Result<[u8; N]> {
+        let mut buffer = [0u8; N];
+        r.read_exact(&mut buffer)?;
+        Ok(buffer)
+    }
+
+    fn is_eof<T>(r: &io::Result<T>) -> bool {
+        if let Err(ref e) = r {
+            if e.kind() == ErrorKind::UnexpectedEof {
+                return true;
+            }
+        }
+        false
+    }
+
+    pub fn unmarshall<R: Read>(r: &mut R) -> io::Result<Self> {
+        let mut slf = Self::default();
+        loop {
+            let res = Self::read_bytes(r);
+            if Self::is_eof(&res) {
+                break;
+            }
+            let k = u32::from_le_bytes(res?);
+
+            let res = Self::read_bytes(r);
+            if Self::is_eof(&res) {
+                break;
+            }
+            let hash = Hash::from(res?);
+
+            let res = Self::read_bytes(r);
+            if Self::is_eof(&res) {
+                break;
+            }
+            let p = u32::from_le_bytes(res?);
+            slf.tree_pos.insert(k, (hash, p as u64));
+        }
+        Ok(slf)
+    }
+
+    pub fn iter(&self) -> impl Iterator<Item = (&u32, &(Hash, u64))> {
+        self.tree_pos.iter()
+    }
+}
+
 #[derive(Debug, Clone, Archive, Deserialize, Serialize)]
 #[archive_attr(derive(CheckBytes))]
 pub struct ContractIndexElement {
-    pub tree: PageTree,
-    pub len: usize,
-    pub page_indices: BTreeSet<usize>,
-    pub hash: Option<Hash>,
-    pub int_pos: Option<u64>,
+    tree: PageTree,
+    len: usize,
+    page_indices: BTreeSet<usize>,
+    hash: Option<Hash>,
+    int_pos: Option<u64>,
+}
+
+impl ContractIndexElement {
+    pub fn new(is_64: bool) -> Self {
+        Self {
+            tree: PageTree::new(is_64),
+            len: 0,
+            page_indices: BTreeSet::new(),
+            hash: None,
+            int_pos: None,
+        }
+    }
+
+    pub fn page_indices_and_tree(
+        self,
+    ) -> (impl Iterator<Item = usize>, PageTree) {
+        (self.page_indices.into_iter(), self.tree)
+    }
+
+    pub fn page_indices(&self) -> &BTreeSet<usize> {
+        &self.page_indices
+    }
+
+    pub fn set_len(&mut self, len: usize) {
+        self.len = len;
+    }
+
+    pub fn len(&self) -> usize {
+        self.len
+    }
+
+    pub fn set_hash(&mut self, hash: Option<Hash>) {
+        self.hash = hash;
+    }
+
+    pub fn hash(&self) -> Option<Hash> {
+        self.hash
+    }
+
+    pub fn set_int_pos(&mut self, int_pos: Option<u64>) {
+        self.int_pos = int_pos;
+    }
+
+    pub fn int_pos(&self) -> Option<u64> {
+        self.int_pos
+    }
+
+    pub fn tree(&self) -> &PageTree {
+        &self.tree
+    }
+
+    pub fn insert_page_index_hash(
+        &mut self,
+        page_index: usize,
+        page_index_u64: u64,
+        page_hash: impl Into<Hash>,
+    ) {
+        self.page_indices.insert(page_index);
+        self.tree.insert(page_index_u64, page_hash);
+    }
 }
 
 impl Default for NewContractIndex {
@@ -191,18 +342,13 @@ impl NewContractIndex {
         self.inner_contracts.insert(*contract_id, element);
     }
 
-    pub fn get(
-        &self,
-        contract: &ContractId,
-        _maybe_commit_id: Option<Hash>,
-    ) -> Option<&ContractIndexElement> {
+    pub fn get(&self, contract: &ContractId) -> Option<&ContractIndexElement> {
         self.inner_contracts.get(contract)
     }
 
     pub fn get_mut(
         &mut self,
         contract: &ContractId,
-        _maybe_commit_id: Option<Hash>,
     ) -> Option<&mut ContractIndexElement> {
         self.inner_contracts.get_mut(contract)
     }
@@ -215,6 +361,12 @@ impl NewContractIndex {
         &self,
     ) -> impl Iterator<Item = (&ContractId, &ContractIndexElement)> {
         self.inner_contracts.iter()
+    }
+
+    pub fn move_into(self, target: &mut Self) {
+        for (contract_id, element) in self.inner_contracts.into_iter() {
+            target.insert_contract_index(&contract_id, element);
+        }
     }
 }
 
@@ -398,4 +550,30 @@ pub fn position_from_contract(contract: &ContractId) -> u64 {
         .fold(0, u32::wrapping_add);
 
     pos as u64
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::io::{BufReader, BufWriter};
+
+    #[test]
+    fn merkle_position_serialization() -> Result<(), io::Error> {
+        const TEST_SIZE: u32 = 262144;
+        const ELEM_SIZE: usize = 4 + 32 + 4;
+        let mut marshalled = TreePos::default();
+        let h = Hash::from([1u8; 32]);
+        for i in 0..TEST_SIZE {
+            marshalled.insert(i, (h, i as u64));
+        }
+        let v: Vec<u8> = Vec::new();
+        let mut w = BufWriter::with_capacity(TEST_SIZE as usize * ELEM_SIZE, v);
+        marshalled.marshall(&mut w)?;
+        let mut r = BufReader::new(w.buffer());
+        let unmarshalled = TreePos::unmarshall(&mut r)?;
+        for i in 0..TEST_SIZE {
+            assert_eq!(unmarshalled.tree_pos.get(&i), Some(&(h, i as u64)));
+        }
+        Ok(())
+    }
 }